telegram_bot/requirements.txt
python-telegram-bot>=20.0
pdfplumber
reportlab
langchain_ollama
emojipy
transformers
torch
six